Safari-611.1.10/LayoutTests/inspector/worker/dom-debugger-event-animation-frame-breakpoints-expected.txt

Tests for animation frame event breakpoint functionality (DOMDebugger) in a Worker.


== Running test suite: Worker.DOMDebugger.EventBreakpoint.AnimationFrame
-- Running test case: Worker.DOMDebugger.EventBreakpoint.AnimationFrame.setEventBreakpoint
PASS: Should produce an exception.
Error: Not supported

-- Running test case: Worker.DOMDebugger.EventBreakpoint.AnimationFrame.removeEventBreakpoint
PASS: Should produce an exception.
Error: Not supported
